fix serialization

+++ core/domain/src/main/java/com/example/juicr/repositories/IPreparationHistory.kt
@@ -4,7 +4,7 @@ import com.example.juicr.models.PreparationHistory
 import kotlinx.coroutines.flow.Flow
 
 interface IPreparationHistory {
-    suspend fun addHistory(prepHistory: PreparationHistory): Int
+    suspend fun addHistory(prepHistory: PreparationHistory): Long
 
     fun getHistoryByJuiceId(id: Long): Flow<List<PreparationHistory>>
     fun getRecentHistory(limit: Int?): Flow<List<PreparationHistory>>
